requirements.txt
requests
Pillow
pycryptodome
